refactor: move all FormsPage child components to it's folder
