+++ react-components/src/components/pages/FormPage/components/ConfirmationWindow.tsx
