+++ react-components/src/components/pages/FormPage/functions/showCreateCardConfirmation.ts
