+++ react-components/src/components/pages/FormPage/functions/toggleSubmitFunctions.ts
