+++ react-components/src/components/pages/FormPage/inputs/AvatarInput.tsx
